requirements.txt
requests>=2.32.0
rich>=13.7.1
botasaurus>=4.0.8
pyjwt>=2.9.0
python-dotenv>=1.0.1
